% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDrug.R
\name{fetchDrugbank}
\alias{fetchDrugbank}
\title{Load Drugbank data (downloaded January 2020)}
\usage{
fetchDrugbank()
}
\value{
dataframe
}
\description{
Load Drugbank data (downloaded January 2020)
}
\examples{
fetchDrugbank()
}
