% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome_subset.R
\docType{data}
\name{interactome_subset}
\alias{interactome_subset}
\title{Interactome data, containing gene-gene, gene-drug associations}
\format{A data frame with 464422 rows and 2 columns
\describe{
 \item{Protein_A}{First entity}
 \item{Protein_B}{Second entity}
}}
\usage{
interactome_subset
}
\description{
Interactome data, containing gene-gene, gene-drug associations
}
\keyword{datasets}
