% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_interactome.R
\name{load_interactome}
\alias{load_interactome}
\title{Loads the interactome information from data}
\usage{
load_interactome(DISEASE1, DISEASE2, DRUG1, DRUG2)
}
\arguments{
\item{DRUG2}{}
}
\value{
list
}
\description{
Loads the interactome information from data
}
