% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_separation.R
\name{measure_separation}
\alias{measure_separation}
\title{Measure separation of two entities (disease or drugs)}
\usage{
measure_separation(diseaseA, diseaseB, gr)
}
\arguments{
\item{gr}{}
}
\value{
data.frame
}
\description{
Measure separation of two entities (disease or drugs)
}
\examples{
library(drugdealR)
load_interactome()

To calculate separation between two drugs, insert drug names
measure_separation('Imatinib', 'Tandutinib', gr)

To calculate separation between two diseases, insert disease names
measure_separation('Myeloid Leukemia, Chronic', 'Multiple Sclerosis')
}
