% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_interactome.R
\name{load_interactome}
\alias{load_interactome}
\title{Loads the protein-protein interaction network}
\usage{
load_interactome()
}
\value{
graph
}
\description{
Loads the protein-protein interaction network
}
\examples{
\dontrun{
library(drugdealR)
load_interactome()
}
}
