% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_interactome.R
\name{load_interactome}
\alias{load_interactome}
\title{Loads the interactome information from data}
\usage{
load_interactome()
}
\value{
dataframe
}
\description{
Loads the interactome information from data
}
\examples{
library(drugdealR)
load_interactome()
}
