% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{measure_dists}
\alias{measure_dists}
\title{Check if distance matrix of a disease changes after removing a given drug}
\usage{
measure_dists(disease, drug, gr, subnet)
}
\arguments{
\item{disease}{name of the disease of interest (compatible with DisNor).}

\item{drug}{name of the drug of interest (compatible with Drugbank).}

\item{gr}{the interactome igraph graph object.}

\item{subnet}{the subnetwork igraph graph object. Contains the disease genes
and drug targets of diseases and drugs of interest.}
}
\value{
logical
}
\description{
Check if distance matrix of a disease changes after removing a given drug
}
\examples{
\dontrun{
library(drugdealR)
measure_dists('Myeloid Leukemia, Chronic', 'Imatinib', gr, subnet)
}
}
