% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_separation.R
\name{measure_separation}
\alias{measure_separation}
\title{Measure separation of two entities (disease or drugs)}
\usage{
measure_separation(genesA, genesB, gr)
}
\arguments{
\item{genesA}{list of genes associated with the given drug/disease.}

\item{genesB}{list of genes associated with the given drug/disease.}

\item{gr}{the subnetwork igraph graph object that contains all genes involved in
diseases (according to DisNor) and drug targets (according to Drugbank) of
interest.}
}
\value{
numeric
}
\description{
Measure separation of two entities (disease or drugs)
}
\examples{
\dontrun{
library(drugdealR)
d <- load_interactome()

To calculate separation between two drugs, insert drug names
measure_separation(d$drug1_genes, d$drug2_genes, gr)

To calculate separation between two diseases, insert disease names
measure_separation(d$disease1_genes, d$disease2_genes, gr)
}
}
