% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_separation.R
\name{measure_separation}
\alias{measure_separation}
\title{Measure separation of two entities (disease or drugs)}
\usage{
measure_separation(entityA, entityB, gr)
}
\arguments{
\item{entityA}{name of the first item to measure separation. Should be compatible
with DisNor (https://disnor.uniroma2.it/) disease names. If is a drug,
should be compatible with Drugbank names.}

\item{entityB}{name of the second item to measure separation. Should be compatible
with DisNor (https://disnor.uniroma2.it/) disease names if is a disease. If is a drug,
should be compatible with Drugbank names.}

\item{gr}{the subnetwork igraph object that contains all genes involved in
diseases (according to DisNor) and drug targets (according to Drugbank) of
interest.}
}
\value{
numeric
}
\description{
Measure separation of two entities (disease or drugs)
}
\examples{
\dontrun{
library(drugdealR)
load_interactome()

To calculate separation between two drugs, insert drug names
measure_separation('Imatinib', 'Tandutinib', gr)

To calculate separation between two diseases, insert disease names
measure_separation('Myeloid Leukemia, Chronic', 'Multiple Sclerosis')
}
}
